\name{kater-package}
\alias{kater-package}
\alias{kater}
\docType{package}
\title{
\packageTitle{kater}
}
\description{
\packageDescription{kater}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kater}
\packageIndices{kater}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{kater}

Maintainer: \packageMaintainer{kater}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
